# Use public conversion methods to avoid reflection