@smarterclayton 2015-12-24
Pretty big wins in conversion

```
benchmark                                      old ns/op     new ns/op     delta
BenchmarkPodConversion-8                       29842         13867         -53.53%
BenchmarkNodeConversion-8                      34235         11467         -66.51%
BenchmarkReplicationControllerConversion-8     25301         13996         -44.68%
BenchmarkPodCopy-8                             4242          4225          -0.40%
BenchmarkNodeCopy-8                            4836          4811          -0.52%
BenchmarkReplicationControllerCopy-8           3968          3884          -2.12%
BenchmarkEncodeCodec-8                         28026         28264         +0.85%
BenchmarkDecodeCodec-8                         85519         73098         -14.52%
BenchmarkDecodeIntoCodec-8                     213909        201374        -5.86%

benchmark                                      old allocs     new allocs     delta
BenchmarkPodConversion-8                       161            81             -49.69%
BenchmarkNodeConversion-8                      196            74             -62.24%
BenchmarkReplicationControllerConversion-8     139            79             -43.17%
BenchmarkPodCopy-8                             25             25             +0.00%
BenchmarkNodeCopy-8                            24             24             +0.00%
BenchmarkReplicationControllerCopy-8           22             22             +0.00%
BenchmarkEncodeCodec-8                         56             56             +0.00%
BenchmarkDecodeCodec-8                         347            299            -13.83%
BenchmarkDecodeIntoCodec-8                     1039           1026           -1.25%

benchmark                                      old bytes     new bytes     delta
BenchmarkPodConversion-8                       12757         9011          -29.36%
BenchmarkNodeConversion-8                      13125         7267          -44.63%
BenchmarkReplicationControllerConversion-8     11829         9235          -21.93%
BenchmarkPodCopy-8                             3280          3280          +0.00%
BenchmarkNodeCopy-8                            3056          3056          +0.00%
BenchmarkReplicationControllerCopy-8           2960          2960          +0.00%
BenchmarkEncodeCodec-8                         5224          5223          -0.02%
BenchmarkDecodeCodec-8                         17679         15374         -13.04%
BenchmarkDecodeIntoCodec-8                     32991         32783         -0.63%
```