@wojtek-t 2015-12-28
It generally LGTM. I added one comment. Feel free to self-apply lgtm label after applying comment and rebasing. Thanks!

BTW - I will be leaving for paternity leave in O(days) for about a month, so I can't guarantee any responsiveness around that time...