@smarterclayton 2015-12-28
Applied comments and rebased, also print out when an inefficient reflection conversion is being used.