@wojtek-t 2015-12-29
@k8s-bot unit test this please