@wojtek-t 2015-12-29
@smarterclayton - please rebase