@wojtek-t 2015-12-29
@smarterclayton - there is something wrong with the PR (I can't see the diff - there were also some other PRs like that, so it's not your fault). I guess it was just a rebase, so please reapply the lgtm label (I can't see the diff)