@smarterclayton 2015-12-29
It was just a trivial rebase - new conversion method got generated and conflicted because of the existing generation.